Json example
This is an example script to build a database for the contacts. It uses an object to generate a dictionary which is then saved as a json file. The SearchFile class can be used to search for keywords in a dictionary and returns a dic with the keyword in it.